requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asttokens==2.4.1
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipython==8.12.3
jedi==0.19.1
jiter==0.7.0
matplotlib-inline==0.1.7
numpy==1.24.4
openai==1.54.0
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pybase64==1.4.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Python_IO==0.3
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
